Turtle game OSTL Backup Project in PYTHON
This was something i build for fun. 
It might not be good enough to be an official presentation.
But the important thing was it helped me to get a hold on python language and more importantly learn pygame.
This is one of those things where i enjoyed making it but it might not seem impressive to the viewer!